renamed the article update component to article editor also fixed the get request for it to be able to update the information to next action put
