+++ src/components/ArticleEditor/ArticleEditor.js
@@ -6,23 +6,38 @@ import {
   fetchSingleArticle,
   updateSingleArticle,
 } from "../../redux/actions/articlesActions";
+import axiosWithAuth from "../../utils/axiosWithAuth";
 
 import "./ArticleEditor.scss";
 
-const ArticleEditor = (props) => {
+const ArticleUpdate = ({ id, title, desc, url, ...props }) => {
   const [editArticle, setEditArticle] = useState({
-    title: props.articleTitle,
-    desc: props.articleDesc,
-    categoryID: props.category,
-    url: props.url,
+    articleID: id,
+    title: title,
+    desc: desc,
+    categoryID: 1,
+    url: url,
   });
+
+  const articleID = props.match.params.id;
+
+  useEffect(() => {
+    console.log("article id:", articleID);
+
+    fetchSingleArticle(editArticle.articleID);
+    axiosWithAuth()
+      .get(`/api/articles/${articleID}`)
+      .then((res) => {
+        setEditArticle(res.data[0]);
+        console.log("Single Article ---> ", props.articles);
+      })
+      .catch((err) => console.error(err.response));
+  }, []);
 
   console.log("editing article here:", editArticle);
 
-  // const history = useHistory();
-
   useEffect(() => {
-    fetchSingleArticle(props.article.id);
+    fetchSingleArticle(id);
   }, []);
 
   const handleChange = (e) => {
@@ -33,51 +48,42 @@ const ArticleEditor = (props) => {
   };
 
   return (
-    <div className="edit-article">
-      <h2>Edit Your Article</h2>
-      <form onSubmit={updateSingleArticle}>
+    <div className="new-article">
+      <form onSubmit={updateSingleArticle(articleID, editArticle)}>
+        <h2>Edit Your Article</h2>
         <input
           type="text"
-          name="articleTitle"
-          value={props.articleTitle}
+          name="title"
+          value={editArticle.title}
           onChange={handleChange}
           placeholder="Title..."
         />
         <input
           type="text"
-          name="articleDesc"
-          value={props.articleDesc}
+          name="desc"
+          value={editArticle.desc}
           onChange={handleChange}
           placeholder="Description..."
         />
         <input
           type="text"
-          name="category"
-          value={props.category}
-          onChange={handleChange}
-          placeholder="Category..."
-        />
-        <input
-          type="text"
           name="url"
-          value={props.url}
+          value={editArticle.url}
           onChange={handleChange}
           placeholder="URL..."
         />
-
-        <button>Add Article</button>
+        <button>Update Article</button>
       </form>
     </div>
   );
 };
 
-const mapStateToProps = (state) => {
+const mapDispatchToProps = () => {
   return {
-    article: state.articlesReducer.data,
+    fetchSingleArticle,
   };
 };
 
-export default connect(mapStateToProps, {
-  fetchSingleArticle,
-  updateSingleArticle,
-})(ArticleEditor);
+export default connect(null, {
+  mapDispatchToProps,
+})(ArticleUpdate);
