--- src/components/ArticleUpdate/ArticleUpdate.scss
@@ -1,4 +1,0 @@
-h2 {
-    text-align: center;
-    font-size: 3rem;
-}
